% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_conversion.R
\name{time_conversion}
\alias{time_conversion}
\title{Time Conversion}
\usage{
time_conversion(years, months)
}
\arguments{
\item{years}{A list containing desired years entered as text elements ie the string "2012"}

\item{months}{A list containing desired months as text elements of the form MM (01-12) ie the string "01" represents January and so on.}
}
\value{
a list of dates formated YYYYMM
}
\description{
Function to take in years and months desired and output a list in the form of YYYYMM
}
\examples{
  #Outputs a list of all first quarter months 2010 - 2020 formatted YYYYMM
  time_conversion(years = list("2010","2011", "2012", "2013", "2014", "2015", "2016", "2017", "2018", "2019", "2020") months = list("01", "02", "03"))
}
